mpos_model_training/requirements.txt
numpy>=1.19.0
pandas>=1.2.0
matplotlib>=3.3.0
scipy>=1.6.0
logomaker>=0.8
tensorflow>=2.5.0
pyyaml>=5.4.0
pulp==2.7  # Pin to 2.7 for Snakemake compatibility (2.8+ uses listSolvers instead of list_solvers)
snakemake>=6.0.0

